app/requirements.txt
boto3==1.24.91
botocore==1.27.91
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
